pyproject.toml
[tool.poetry]
name = "r2r-scrapy"
version = "1.0.0"
description = "Asynchronous Documentation Collector for RAG Systems"
authors = ["Evgeny Gurin <e.a.gurin@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/eagurin/r2r-scrapy"
repository = "https://github.com/eagurin/r2r-scrapy"
documentation = "https://github.com/eagurin/r2r-scrapy/docs"
keywords = ["scrapy", "rag", "documentation", "scraping", "async"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: Scrapy",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.9"
scrapy = "^2.8.0"
aiohttp = "^3.8.4"
asyncio = "^3.4.3"
pyyaml = "^6.0"
click = "^8.1.3"
tqdm = "^4.65.0"
langchain = "^0.0.2"
langchain-text-splitters = "^0.0.1"
beautifulsoup4 = "^4.12.2"
markdown = "^3.4.3"
html2text = "^2020.1.16"
nbformat = "^5.9.0"
openai = "^0.27.8"
tiktoken = "^0.4.0"
cryptography = "^41.0.3"
prometheus-client = "^0.17.1"
psutil = "^5.9.5"
gitpython = "^3.1.32"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
pygments = "^2.15.0"
playwright = "^1.36.0"

[tool.poetry.group.js]
optional = true

[tool.poetry.group.js.dependencies]
scrapy-splash = "^0.8.0"
playwright = "^1.36.0"

[tool.poetry.group.distributed]
optional = true

[tool.poetry.group.distributed.dependencies]
redis = "^4.6.0"
scrapy-redis = "^0.7.3"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
spacy = "^3.5.3"
nltk = "^3.8.1"

[tool.poetry.group.openapi]
optional = true

[tool.poetry.group.openapi.dependencies]
jsonschema = "^4.17.3"

[tool.poetry.group.translate]
optional = true

[tool.poetry.group.translate.dependencies]
deepl = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
ruff = "^0.0.280"
mypy = "^1.5.1"
sphinx = "^7.1.2"
sphinx-rtd-theme = "^1.3.0"
pre-commit = "^3.3.3"

[tool.poetry.scripts]
r2r-scrapy = "r2r_scrapy.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true 